src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/title-element-in-shadow-expected.txt

CONSOLE ERROR: line 21: Uncaught TypeError: subtree1.createShadowRoot is not a function
This test ensures that title elements in a shadow subtree do not affect document.title attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

create a title element and insert it to document.
PASS document.title is "document title"
remove title element from document.
PASS document.title is ""
create a shadow root whose host is already in document, create a subtree which contains a title element, and add the subtree to the shadow root.
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
